condor/pyproject.toml
[project]
name = "condor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "htcondor>=24.8.1",
    "ipython>=9.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
